insert anchors for indirect references without sectionsplit: https://github.com/metanorma/metanorma/issues/373
