+++ spec/collection_spec.rb
@@ -166,12 +166,11 @@ RSpec.describe Metanorma::Collection do
       expect(File.exist?("#{OUTPATH}/rice1-en.final.presentation.xml"))
         .to be true
       rice = File.read("#{OUTPATH}/rice-en.final.html")
-      require "debug"; binding.b
       expect(rice).to include %(This document is updated in <a href="rice-amd.final.html"><span class="stdpublisher">ISO </span><span class="stddocNumber">17301</span>-<span class="stddocPartNumber">1</span>:<span class="stdyear">2016</span>/Amd.1:2017</a>.</p>)
       expect(rice).to include %(It is not applicable to cooked rice products, which are not discussed in <a href="#anotherclause_ISO_17301-1_2016"><span class="citesec">Clause 2</span></a> or <a href="#thirdclause_ISO_17301-1_2016"><span class="citesec">Clause 3</span></a>.</p>)
       # demonstrate that erefs are removed if they point to another document in the repository,
       # but that document is not supplied
-      expect(rice).to include %{This document uses schemas E0/A0, <a href="example/url.html#A1">E1/A1</a> and <a href="example/url.html#A2">E2/A2</a>.}
+      expect(rice).to include %{This document uses schemas E0/A0, <a href="example/url.html#A1">E1/A1</a> and <a href="example/url.html#E2">E2</a>.}
       expect(rice).to include %(This document is also unrelated to <a href="example/url.html#what">)
       FileUtils.rm_rf of
     end
@@ -413,7 +412,7 @@ RSpec.describe Metanorma::Collection do
       expect(rice).to include %(It is not applicable to cooked rice products, which are not discussed in <a href="rice-en.final.xml.2.html#anotherclause_ISO_17301-1_2016_ISO_17301-1_2016_2_This_is_another_clause"><span class="citesec">Clause 2</span></a> or <a href="rice-en.final.xml.3.html#thirdclause_ISO_17301-1_2016_ISO_17301-1_2016_3_This_is_another_clause"><span class="citesec">Clause 3</span></a>.</p>)
       # demonstrate that erefs are removed if they point to another document in the repository,
       # but that document is not supplied
-      expect(rice).to match %r{This document uses schemas E0/A0, <a href="dummy.html#express-schema_E1_ISO_17302">E1/A1</a> and <a href="dummy.html#express-schema_E2_ISO_17302">E2/A2</a>.}
+      expect(rice).to match %r{This document uses schemas E0/A0, <a href="dummy.html#express-schema_E1_ISO_17302">E1/A1</a> and <a href="dummy.html#express-schema_E2_ISO_17302">express-schema/E2</a>.}
       expect(rice).to include %(This document is also unrelated to <a href="dummy.html#what">)
       xml = Nokogiri::XML(File.read("#{OUTPATH}/rice-en.final.xml.1.presentation.xml"))
       p = xml.xpath("//xmlns:sections//xmlns:p")[4]
@@ -476,7 +475,7 @@ RSpec.describe Metanorma::Collection do
       expect(File.read("#{OUTPATH}/rice-en.final.html"))
         .to include %(This document is also unrelated to <a href="dummy.xml.3.html#what">)
       expect(File.read("#{OUTPATH}/rice-en.final.html"))
-        .to match %r{This document uses schemas E0/A0, <a href="dummy.xml.2.html#A1_ISO_17302_ISO_17302_2">E1/A1</a> and <a href="dummy.xml.4.html#A2_ISO_17302_ISO_17302_4">E2/A2</a>.</p>}
+        .to include %{This document is also unrelated to <a href="dummy.xml.3.html#what">current-metanorma-collection/ISO 17302 3 What?</a>.</p><p id="_001_ISO_17301-1_2016">This document uses schemas E0/A0, <a href="dummy.xml.2.html#A1_ISO_17302_ISO_17302_2">E1/A1</a> and <a href="dummy.xml.4.html#E2_ISO_17302_ISO_17302_4">E2</a>.</p>}
       FileUtils.rm_rf of
     end
