+++ spec/fixtures/collection/rice-en.final.xml
@@ -103,7 +103,8 @@ at the address below or ISO’s member body in the country of the requester.
 <p id="_b9d7b0c0-ff21-4205-b5b5-8435634a0c95">This document is updated in <eref type="inline" bibitemid="RiceAmd" citeas="ISO 17301-1:2016/Amd.1:2017"/>.</p>
 <p id="_50a0e498-4b20-401c-ab99-fed8d11e266b">This document is unrelated <eref type="inline" bibitemid="StarTrek" citeas="ISO 1701:1974"/>, see e.g. <eref type="inline" bibitemid="StarTrek" citeas="ISO 1701:1974"><localityStack><locality type="clause"><referenceFrom>1</referenceFrom></locality></localityStack></eref>.</p>
 <p id="_60bc26a9-2eaa-4f54-97f6-b58af731babb">This document is also unrelated to <eref type="inline" bibitemid="Dummy" citeas="ISO 17302"><localityStack><locality type="anchor"><referenceFrom>what</referenceFrom></locality></localityStack></eref>.</p>
-<p id="_001">This document uses schemas <eref bibitemid="express-schema_E0"><locality type="anchor"><referenceFrom>A0</referenceFrom></locality>E0/A0</eref>, <eref bibitemid="express-schema_E1"><locality type="anchor"><referenceFrom>A1</referenceFrom></locality>E1/A1</eref> and <eref bibitemid="express-schema_E2"><locality type="anchor"><referenceFrom>A2</referenceFrom></locality>E2/A2</eref>.</p>
+<!--<p id="_001">This document uses schemas <eref bibitemid="express-schema_E0"><locality type="anchor"><referenceFrom>A0</referenceFrom></locality>E0/A0</eref>, <eref bibitemid="express-schema_E1"><locality type="anchor"><referenceFrom>A1</referenceFrom></locality>E1/A1</eref> and <eref bibitemid="express-schema_E2"><locality type="anchor"><referenceFrom>A2</referenceFrom></locality>E2/A2</eref>.</p>-->
+<p id="_001">This document uses schemas <eref bibitemid="express-schema_E0"><locality type="anchor"><referenceFrom>A0</referenceFrom></locality>E0/A0</eref>, <eref bibitemid="express-schema_E1"><locality type="anchor"><referenceFrom>A1</referenceFrom></locality>E1/A1</eref> and <eref bibitemid="express-schema_E2"></eref>.</p>
 <svgmap id='svg1'>
   <figure id='fig1'>
     <image src='action_schemaexpg1.svg' id='_' mimetype='image/svg+xml' height='auto' width='auto' alt='Workmap'/>
